copy_trading_backend/src/test/java/com/xtq_ymt/copy_trading_backend/service/StatsServiceImplTest.java
package com.xtq_ymt.copy_trading_backend.service;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.mockito.*;

import static org.mockito.Mockito.*;
import static org.junit.jupiter.api.Assertions.*;

/**
 * 单元测试类：StatsServiceImplTest
 * 负责测试 StatsServiceImpl 类的方法，确保其行为符合预期。
 */
public class StatsServiceImplTest {

    @Mock
    private TradeService tradeService;  // 模拟 TradeService 接口

    @InjectMocks
    private StatsServiceImpl statsService;  // 被测试的 StatsServiceImpl 实例

    /**
     * 在每个测试方法执行之前都会调用这个方法，初始化测试所需的 mock 对象。
     * 使用 MockitoAnnotations.openMocks(this) 手动初始化 mock 对象。
     */
    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);  // 手动初始化 mock 对象
    }

    /**
     * 测试 StatsServiceImpl 的 updateTraderAndFollowerStats 方法。
     * 确保当调用该方法时，会正确调用 TradeService 的相应方法。
     */
    @Test
    @DisplayName("测试 updateTraderAndFollowerStats 方法调用 TradeService 的方法")
    void testUpdateTraderAndFollowerStats() {
        Long tradeId = 1L;  // 模拟传入的交易 ID

        // 设置 mock 对象的行为：当调用 tradeService.updateTraderAndFollowerStats(tradeId) 时，什么也不做
        doNothing().when(tradeService).updateTraderAndFollowerStats(tradeId);

        // 调用 StatsServiceImpl 的方法，实际上会触发对 TradeService 的调用
        statsService.updateTraderAndFollowerStats(tradeId);

        // 验证 tradeService 的 updateTraderAndFollowerStats 方法被调用了一次
        verify(tradeService, times(1)).updateTraderAndFollowerStats(tradeId);
    }

    /**
     * 测试 StatsServiceImpl 的 updateTraderAndFollowerStats 方法在 TradeService 抛出异常时的行为。
     * 确保异常能够被正确传播。
     */
    @Test
    @DisplayName("测试 updateTraderAndFollowerStats 方法在 TradeService 抛出异常时的行为")
    void testUpdateTraderAndFollowerStats_Exception() {
        Long tradeId = 2L;  // 模拟传入的交易 ID

        // 设置 mock 对象的行为：当调用 tradeService.updateTraderAndFollowerStats(tradeId) 时，抛出异常
        doThrow(new IllegalArgumentException("Trade not found")).when(tradeService).updateTraderAndFollowerStats(tradeId);

        // 调用 StatsServiceImpl 的方法，并断言抛出相应的异常
        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
            statsService.updateTraderAndFollowerStats(tradeId);
        });

        // 验证异常消息是否正确
        assertEquals("Trade not found", exception.getMessage());

        // 验证 tradeService 的 updateTraderAndFollowerStats 方法被调用了一次
        verify(tradeService, times(1)).updateTraderAndFollowerStats(tradeId);
    }

    /**
     * 示例测试方法：假设 StatsServiceImpl 有一个获取 TraderStats 的方法。
     * 根据您的具体实现添加相关的测试逻辑。
     */
    @Test
    @DisplayName("测试获取 TraderStats 方法")
    void testGetTraderStats() {
        // 根据您的具体实现添加相关的测试逻辑
        // 例如，如果 StatsServiceImpl 有一个 getTraderStats(Long traderId) 方法：

        /*
        Long traderId = 1L;
        TraderStats mockStats = new TraderStats();
        mockStats.setTraderId(traderId);
        mockStats.setTotalProfit(1000.0);
        mockStats.setWinRate(75.0);
        mockStats.setMaxDrawdown(-500.0);
        mockStats.setTotalTrades(20);
        mockStats.setWinningTrades(15);
        mockStats.setTotalFollowers(10);
        mockStats.setUser(new User(traderId, "trader1", User.Role.TRADER));

        when(tradeService.getTraderStats(traderId)).thenReturn(mockStats);

        TraderStats stats = statsService.getTraderStats(traderId);

        assertNotNull(stats);
        assertEquals(1000.0, stats.getTotalProfit());
        assertEquals(75.0, stats.getWinRate());
        assertEquals(-500.0, stats.getMaxDrawdown());
        assertEquals(20, stats.getTotalTrades());
        assertEquals(15, stats.getWinningTrades());
        assertEquals(10, stats.getTotalFollowers());
        assertEquals("trader1", stats.getUsername());

        verify(tradeService, times(1)).getTraderStats(traderId);
        */
    }
}
